move UniqueId
